+++ app/uk/gov/hmrc/tai/connectors/HipHeaders.scala
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2025 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.connectors
+
+import play.api.http.MimeTypes
+import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames}
+
+import java.nio.charset.StandardCharsets
+import java.util.{Base64, UUID}
+
+object HipHeaders {
+  def get(
+    originatorId: String,
+    hc: HeaderCarrier,
+    hipExtraInfo: Option[(String, String)]
+  ): Seq[(String, String)] = {
+    val hipAuth = hipExtraInfo.fold[Seq[(String, String)]](Seq.empty) { case (clientId, clientSecret) =>
+      val token = Base64.getEncoder.encodeToString(s"$clientId:$clientSecret".getBytes(StandardCharsets.UTF_8))
+      Seq(
+        HeaderNames.authorisation -> s"Basic $token"
+      )
+    }
+    Seq(
+      play.api.http.HeaderNames.CONTENT_TYPE -> MimeTypes.JSON,
+      "Gov-Uk-Originator-Id"                 -> originatorId,
+      HeaderNames.xSessionId                 -> hc.sessionId.fold("-")(_.value),
+      HeaderNames.xRequestId                 -> hc.requestId.fold("-")(_.value),
+      "CorrelationId"                        -> UUID.randomUUID().toString
+    ) ++ hipAuth
+  }
+}
